CORS: Take TryInto instead of TryFrom
This allows usage of custom types that implement TryInto for the
affected methods, something that was not possible with TryFrom due to
the orphan rule.
As with Into, the standard library provides a blanket implementation of
TryInto<U> for T if U implements TryFrom<T>, making this a fully
backwards-compatible change.
It also allows for simplification of internal code in some places, as it
permits turning nested function calls into method chains.