Corrected actix-form-data community crate details
Next to the actix-form-data community crate you have it listed as "Rate-limiting backed by form-data." from reading the crate description it seems to have nothing to do with rate limiting. Could this be a copy paste error from the crate after it? I feel as though something like what I've replaced it with is more fitting to its purpose/